Create ignored and tracked and vice versa
The first line of the input contains three integers, N. M, and K, denoting the number of source files in the project, the number of ignored source files, and the number of tracked source files. Assume that the maximum value for N is 50. The second line contains M distinct integers denoting the sequence A of ignored source files. The sequence is strictly increasing.

The third line contains K distinct integers denoting the sequence B of tracked source files. The sequence is strictly increasing

Output format:

The output is a single line containing two integers: the number of source files that are both tracked and ignored and the number of source files that are both untracked and unignored.

Sample test cases

Input 1:

746

1467

123467
o1 4 1